requirements.txt
absl-py==0.12.0
asgiref==3.3.4
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
Django==3.2.2
django-crispy-forms==1.11.2
django-filter==2.4.0
djangorestframework==3.12.4
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
imutils==0.5.4
iniconfig==1.1.1
instabot==0.117.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.2
mock==4.0.3
moviepy==1.0.3
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
pluggy==0.13.1
proglog==0.1.9
progressbar2==3.53.1
protobuf==3.16.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.13.3
rsa==4.7.2
schedule==1.1.0
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
svgwrite==1.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc3
tensorflow-estimator==2.5.0rc0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.4.8
toml==0.10.2
tqdm==4.61.1
Tree==0.2.4
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
